producer/requirements.txt
pandas==2.1.1
openpyxl==3.1.2
kafka-python==2.1.0
matplotlib==3.8.0